automate full data update (WIP) - use github acitons to build dataform ontainer (WIP)
