+++ App/YtReader/Db/Snowflake.cs
@@ -5,16 +5,12 @@ using SysExtensions.Security;
 
 namespace YtReader {
   public class SnowflakeCfg {
-    public NameSecret Creds     { get; set; } = new NameSecret();
+    public            NameSecret Creds     { get; set; } = new NameSecret();
     [Required] public string     Account   { get; set; }
     [Required] public string     Warehouse { get; set; }
     [Required] public string     Db        { get; set; }
     [Required] public string     Schema    { get; set; }
-    
-    
   }
-  
-  
 
   public static class SnowflakeEx {
     public static async Task<SnowflakeDbConnection> OpenConnection(this SnowflakeCfg cfg) {
@@ -24,9 +20,9 @@ namespace YtReader {
     }
 
     public static SnowflakeDbConnection Connection(this SnowflakeCfg cfg) => new SnowflakeDbConnection
-      {ConnectionString = cfg.ConnectionStirng(), Password = cfg.Creds.SecureString()};
+      {ConnectionString = cfg.Cs(), Password = cfg.Creds.SecureString()};
 
-    public static string ConnectionStirng(this SnowflakeCfg cfg) =>
+    public static string Cs(this SnowflakeCfg cfg) =>
       $"account={cfg.Account};user={cfg.Creds.Name};db={cfg.Db};schema={cfg.Schema};warehouse={cfg.Warehouse}";
   }
 }
