+++ Dataform/.dockerignore
@@ -1,0 +1,3 @@
+.git
+node_modules/*
+.run
