+++ Dataform/.gitignore
@@ -1,0 +1,3 @@
+.env
+.df-credentials.json
+.run
